# Add stringent ASCII-checks to Sequence object