--- a/src/dnaio/ascii_check.h
+++ b/src/dnaio/ascii_check.h
@@ -7,7 +7,20 @@
 static inline int 
 string_is_ascii(char * string, size_t length) {
     size_t n = length;
-    uint64_t *longword_ptr = (uint64_t *)string;
+    char * char_ptr = string;
+    // The first loop aligns the memory address. Char_ptr is cast to a size_t
+    // to return the memory address. Uint64_t is 8 bytes long, and the processor
+    // handles this better when its address is a multiplier of 8. This loops
+    // handles the first few bytes that are not on such a multiplier boundary
+    // by checking if the 1, 2 or 4 bits are set.
+    while ((size_t)char_ptr & 0b111 && n != 0) {
+        if (*char_ptr & ASCII_MASK_1BYTE) {
+            return 0;
+        }
+        char_ptr += 1;
+        n -= 1;
+    }
+    uint64_t *longword_ptr = (uint64_t *)char_ptr;
     while (n >= sizeof(uint64_t)) {
         if (*longword_ptr & ASCII_MASK_8BYTE){
             return 0;
@@ -15,7 +28,7 @@ string_is_ascii(char * string, size_t length) {
         longword_ptr += 1;
         n -= sizeof(uint64_t);
     }
-    char * char_ptr = (char *)longword_ptr;
+    char_ptr = (char *)longword_ptr;
     while (n != 0) {
         if (*char_ptr & ASCII_MASK_1BYTE) {
             return 0;