Fixes #40 

Only ASCII strings are valid in FASTQ. This PR:
+ Makes the name, sequence and qualities attributes in Sequence internal hidden _name, _sequence, _qualities.
+ Implements custom getters and setters that do check ASCII for Sequence objects
+ Implements ASCII checks in Sequence `__init__` method.
+ Removes all 1BYTE_KIND and similar checks in all of Sequence's methods. Because of the above points, these are guaranteed.
+ Adds an ASCII-check that checks the entire record so that ASCII strings can be safely created for it in `FastqIter.__next__`. This ASCII check is implemented with a C-function.
+ Removes the str typing in `__cinit__`. This removes any PyUnicode_Checks that cython does. We already ensure the entered objects are valid Python Unicode ASCII strings, so there is no need for these checks.

The speed impact is minimal:
Before:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):     912.6 ms ±  16.3 ms    [User: 797.2 ms, System: 115.3 ms]
  Range (min … max):   894.1 ms … 947.5 ms    10 runs

Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.249 s ±  0.017 s    [User: 2.039 s, System: 0.210 s]
  Range (min … max):    2.225 s …  2.277 s    10 runs

Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.542 s ±  0.027 s    [User: 1.432 s, System: 0.110 s]
  Range (min … max):    1.504 s …  1.579 s    10 runs
```

After
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):     945.9 ms ±  12.5 ms    [User: 828.9 ms, System: 116.9 ms]
  Range (min … max):   930.8 ms … 966.4 ms    10 runs

Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.315 s ±  0.015 s    [User: 2.111 s, System: 0.203 s]
  Range (min … max):    2.291 s …  2.328 s    10 runs

Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.600 s ±  0.020 s    [User: 1.473 s, System: 0.126 s]
  Range (min … max):    1.562 s …  1.633 s    10 r
```

The advantage of this PR is that the user can **never** create invalid FASTQ or FASTA records anymore without raising an error.

EDIT: Apparently Cython is not very strict and allows setting `None` for any attribute. So `dnaio.Sequence(None, None, None)` would not raise any errors. That is rectified now. Only qualities may be None.