@marcelm 2022-02-21
While trying to figure out what the performance impact of this actually is, I got derailed and set up airspeed velocity (asv) for dnaio. Here are the benchmark results so far: https://marcelm.github.io/dnaio-asv/. You should have received an invitation to the repository in case you want to contribute a benchmark.

asv can also be used to compare two commits, and for some reason, I get the result that this PR slows down parsing by 10% (presumably just like #41). Here is the output (showing only results for Python 3.10):
```
       before           after         ratio
     [faba8111]       [65885459]
     <main>           <rhpvorderman-asciicheck2>
+        993±20ms       1.16±0.04s     1.17  benchmarks.EmptyRecords.time_parse_empty_records [.../conda-py3.10]
+         397±9ms         461±20ms     1.16  benchmarks.time_parse_single('paired300bp_1') [.../conda-py3.10]
+      1.82±0.02s       2.11±0.02s     1.16  benchmarks.time_parse_single('single43bp') [.../conda-py3.10]
+         436±6ms         504±10ms     1.16  benchmarks.time_parse_single('paired150bp_1') [.../conda-py3.10]
+        970±30ms       1.09±0.02s     1.12  benchmarks.time_parse_paired('paired300bp') [.../conda-py3.10]

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```

I *actually* just wanted to check whether accessing the attributes of the Sequence object is now slower, but unlike in regular Python, that’s not the case. I learned that
- a descriptor implemented in C is very fast
- an attribute declared as `cdef public`/`cdef readonly` in Cython is implemented as a descriptor anyway, so it makes no difference at all (the `Py_tp_getset` slot is filled the same way before and after this PR).

I’ll re-run my benchmarks and maybe a different one. I’m not against some slowdown for the safety this gives, but then it’s probably easier to just use the existing `PyUnicode_...` function for the ASCII check.