@rhpvorderman 2022-02-21
> While trying to figure out what the performance impact of this actually is, I got derailed and set up airspeed velocity (asv) for dnaio. Here are the benchmark results so far: https://marcelm.github.io/dnaio-asv/. You should have received an invitation to the repository in case you want to contribute a benchmark.

Cool! I will check it out.

> I’ll re-run my benchmarks and maybe a different one. I’m not against some slowdown for the safety this gives, but then it’s probably easier to just use the existing PyUnicode_... function for the ASCII check.

On my machine that is slower than our current method. 
Main:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):     916.1 ms ±  17.2 ms    [User: 801.2 ms, System: 114.8 ms]
  Range (min … max):   889.0 ms … 942.3 ms    10 runs
```
This branch:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.056 s ±  0.015 s    [User: 933.4 ms, System: 122.7 ms]
  Range (min … max):    1.029 s …  1.082 s    10 runs
```
Using PyUnicode_DecodeASCII
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.301 s ±  0.057 s    [User: 1.170 s, System: 0.126 s]
  Range (min … max):    1.232 s …  1.417 s    10 runs
```