@rhpvorderman 2022-02-21
> an attribute declared as cdef public/cdef readonly in Cython is implemented as a descriptor anyway, so it makes no difference at all (the Py_tp_getset slot is filled the same way before and after this PR)

Indeed. The getter is just increasing a reference count and returning an existing pointer, so that is as fast as it can be. The setter however, should be a little slower because it has more checks. But that is okay, since most of these checks are C macros (fast), and FASTQ files are typically read and not modified extensively. 

It is a bit unfortunate that Cython implements the setters so nonsensically. Allowing None to be set to every attribute. Very annoying. I had to write tests for that to explicitly make sure I had implemented it correctly.