@rhpvorderman 2022-02-21
I won't do this now, but if we get crazy later, the ascii check can be performed with vectorized instruction using x86 intrinsics. That way 16, 32 or 64 characters can be checked at the same time. It seems like a nice contained use case to use these functions.