@marcelm 2022-02-21
> the ascii check can be performed with vectorized instruction using x86 intrinsics

If you do that, you should submit that as a PR to CPython ...