@rhpvorderman 2022-02-22
> If you do that, you should submit that as a PR to CPython ...

I already tested  it. There is no advantage over using 64-bit integers and 8-byte masks. The instruction that enables the test is [`_mm_testz_si128`](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_testz_si128&ig_expand=7252). But that function internally calculates two values. ZF and CF. So it does two operations where a bit mask does one. I could not measure a difference. I also tried `_mm_testz_si256` but also no difference.

On top of that, this needs special compiler instructions `-msse4.1` for `_mm_testz_si128`. That is OK, there are no x86 cpus that do not support that anymore (2006 or later). But for `_mm_testz_si256` you need `-mavx` and that is quite a bit more modern. 
Plus this won't work for ARM builds and it makes the code more complex, and I simply could not get it to be faster. So it is not worth it.

I did measure the difference between applying an 8-byte mask and a 1-byte mask everywhere and that difference was quite noticable. The parse time was almost doubled using the naive implementation.