@marcelm 2022-02-28
So after redoing some of the benchmarks, the results from https://github.com/marcelm/dnaio/pull/48#issuecomment-1046738257 still stand: The slowdown is about 15% on my machine. (And PyUnicode_DecodeASCII is slower.)

So just as a final confirmation before hitting merge: Do you agree this is worth it?