@rhpvorderman 2022-03-02
Definitely. Outputting ASCII strings is exactly what the user should expect. If they want unsafe and speedy, the binary mode is available now.
Also the current method of doing things is still faster than v0.7.1 and includes an ASCII-check now. (PyUnicode_DecodeLatin1 is also slower...) .